+++ tasty-plutus/src/Test/Tasty/Plutus/Script/Property.hs
@@ -66,21 +66,11 @@ import Test.QuickCheck (
   stdArgs,
  )
 import Test.Tasty.Options (OptionDescription (Option), OptionSet, lookupOption)
-import Test.Tasty.Plutus.Internal (
+import Test.Tasty.Plutus.Internal (ourStyle)
+import Test.Tasty.Plutus.Internal.Context (
   ContextBuilder,
-  PropertyMaxSize (PropertyMaxSize),
-  PropertyTestCount (PropertyTestCount),
   Purpose (ForMinting, ForSpending),
-  ScriptResult (
-    InternalError,
-    NoOutcome,
-    ParseFailed,
-    ScriptFailed,
-    ScriptPassed
-  ),
   TransactionConfig,
-  WithScript (WithMinting, WithSpending),
-  ourStyle,
  )
 import Test.Tasty.Plutus.Internal.Env (
   SomeScript (SomeMinter, SomeSpender),
@@ -99,8 +89,25 @@ import Test.Tasty.Plutus.Internal.Feedback (
   unexpectedFailure,
   unexpectedSuccess,
  )
+import Test.Tasty.Plutus.Internal.Options (
+  PropertyMaxSize (PropertyMaxSize),
+  PropertyTestCount (PropertyTestCount),
+ )
+import Test.Tasty.Plutus.Internal.Run (
+  ScriptResult (
+    InternalError,
+    NoOutcome,
+    ParseFailed,
+    ScriptFailed,
+    ScriptPassed
+  ),
+ )
+import Test.Tasty.Plutus.Internal.WithScript (
+  WithScript (WithMinting, WithSpending),
+ )
 import Test.Tasty.Plutus.Options (
   Fee,
+  ScriptInputPosition,
   TestCurrencySymbol,
   TestTxId,
   TestValidatorHash,
@@ -264,6 +271,7 @@ instance (Typeable p) => IsTest (PropertyTest p) where
       , Option @TestValidatorHash Proxy
       , Option @PropertyTestCount Proxy
       , Option @PropertyMaxSize Proxy
+      , Option @ScriptInputPosition Proxy
       ]
 
 spenderProperty ::