+++ plutus-extra/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 3.0 -- 2021-11-11
+
+### Changed
+  
+* Plutus upgrade: `plutus` pinned to `3f089ccf0ca746b399c99afe51e063b0640af547`, 
+  `plutus-apps` pinned to `404af7ac3e27ebcb218c05f79d9a70ca966407c9`
+
+## 2.2 -- 2021-11-01
+
+### Added
+
+- `valueAtComputedAddressWithState` function at `Plutus.Contract.Test.Extra`
+
+### Changed
+
+* Unified output of predicate builders that use computed address
+
 ## 2.1 -- 2021-10-22
 
 ### Added